requirements.txt
gunicorn
line-bot-sdk
flask
audioread
google-generativeai
pillow
configparser